Open in new tab option, fragments, uri schemes, and fixes
Url checking

I noticed that when you don't add a protocol to a url like this
www.hello.world
the url added to the editor attempts to reference a file
file:///path/to/project/www.hello.world

So I took a look at how to require http or https and noticed
Utils.isUrl() is not working properly in another way
due to the regex.

This is passing the regex
www.0.♪┏(・o・)┛♪

This is failing
http://wwwuzzzuppp.com

I've updated the regex to fix the regex issue and require the protocol.
Adding links should be done through the api or absolute urls.

---

closes #8
Impossible to disable server search Issue #8
https://github.com/editor-js/link-autocomplete/issues/8

isServerEnabled method calls lacked parenthesis in index.js. It's
now possible to disable the server calls by removing the endpoint
in the config.

---

closes #6
Support mailto: URL schema #6
https://github.com/editor-js/link-autocomplete/issues/6

I've found myself wishing to to add different tyes of hrefs and
noticed this issue.

I've added 3 URI schemes to check against in Utils.isUrl()

view-source:
mailto:
tel:

It now checks for a URI Scheme and then applies an appropriate
regex to test the remaining url.

The email regex uses control characters so disabled no-control-regex in .eslintrc

---

Url fragments not recongnized as urls.

Utils.isUrl() now also checks if your string starts with "#"

so you can do this `<a href="#lower_in_the_document">`

Which will work great with this plugin.
https://github.com/Aleksst95/header-with-anchor

---

Can't open links in a new tab.

A checkbox has been added to provide the user an option to open the url in a new tab.
You can tab and shift-tab back and forth between the 2 inputs when adding urls.
Clicking the datawrapper will let you know if the link opens in a new tab or the same tab.

---

Finally, I've fixed a couple warnings that were shown when running the linter.